feat: introduce ExternalAPI
