+++ common/hedera-common-nativesupport/src/main/java/module-info.java
@@ -5,4 +5,5 @@
  */
 module com.hedera.common.nativesupport {
     exports com.hedera.common.nativesupport;
+    exports com.hedera.common.nativesupport.api;
 }
